+++ R/flux_plot_exp.R
@@ -5,6 +5,7 @@
 #' @param f_conc column with gas concentration
 #' @param f_datetime column with datetime of each data point
 #' @param y_text_position position of the text box
+#' @param kappamax indicating if kappamax was applied
 #' @importFrom dplyr select distinct mutate
 #' @importFrom ggplot2 ggplot aes geom_point geom_line theme_bw geom_vline
 #' scale_color_manual scale_x_datetime ylim facet_wrap labs geom_text
@@ -18,20 +19,17 @@
 flux_plot_exp <- function(slopes_df,
                           f_conc,
                           f_datetime,
-                          y_text_position) {
+                          y_text_position,
+                          kappamax) {
 
-  kappamax <- attributes(slopes_df)$kappamax
+
+
+  if (!is.null(kappamax) && kappamax == TRUE) {
+    param_df <- flux_param_kappamax(slopes_df)
+  }
 
   if (is.null(kappamax)) {
-    kappamax <- FALSE
-  }
-
-  if (kappamax == TRUE) {
-    param_df <- flux_param_kappamax(slopes_df, {{f_conc}})
-  }
-
-  if (kappamax == FALSE) {
-    param_df <- flux_param_exp(slopes_df, {{f_conc}})
+    param_df <- flux_param_exp(slopes_df)
   }
 
   slopes_df <- flux_plot_flag(slopes_df, param_df)