+++ R/flux_plot_flag.R
@@ -9,7 +9,7 @@ flux_plot_flag <- function(slopes_df,
                            param_df) {
   slopes_df <- slopes_df |>
     select(!c("f_quality_flag")) |>
-    left_join(param_df, by = "f_fluxid") |>
+    left_join(param_df, by = "f_facetid") |>
     mutate(
       f_quality_flag = case_when(
         f_cut == "cut" ~ f_cut,