+++ building-material/dev.R
@@ -3,7 +3,7 @@ usethis::use_testthat()
 usethis::use_test("flux_fitting_quadratic")
 
 #to add used packages in the description
-usethis::use_package("lifecycle")
+usethis::use_package("tibble")
 
 # to make the webpage
 usethis::use_pkgdown()